packages/sdc-core/pyproject.toml
[project]
name = "sdc-core"
version = "0.1.0"
description = "Shared framework for Social Data Commons dataset pipelines"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pandas>=2.2",
    "click>=8.1",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
geo = ["geopandas>=1.0"]
dev = ["pytest>=8.0", "ruff>=0.8"]

[project.scripts]
sdc = "sdc_core.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sdc_core"]

[tool.ruff]
line-length = 100
